Prepend svg to body instead of append
Currently the library appends the SVG inline to the end of the document body. Some older browsers (for example Safari 6) will not work properly with SVG sprites if you try to do a <use> that references an SVG sprite that is definied AFTER the <use> in the HTML.

http://stackoverflow.com/a/22969530/172350

The solution is to make sure the SVG is inserted at the top of the document, just after the opening <body> tag. That's what this change does. Instead of inserting the SVG at the end of the <body> it puts it at the beginning. Everything works exactly as before but this covers some older browsers that would break otherwise.

Signed-off-by: Jake Wilson <jake.wilson@otterproducts.com>